+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix broken core snapshot tests
 - Improve NixOS support
+- Add line number to failing tests
 
 ## [0.22.1](https://github.com/TypedDevs/bashunit/compare/0.22.0...0.22.1) - 2025-07-23